Fix pre_read & post_read function
