--- .vscode/settings.json
@@ -1,10 +1,0 @@
-{
-    "files.associations": {
-        "nct7363.h": "c",
-        "plat_sensor_table.h": "c",
-        "sensor.h": "c",
-        "ast_adc.h": "c",
-        "hdc1080.h": "c",
-        "tmp461.h": "c"
-    }
-}
