+++ meta-facebook/aalc_predev/src/platform/plat_hook.c
@@ -19,8 +19,17 @@
 #include <logging/log.h>
 #include "plat_hook.h"
 #include "plat_class.h"
-#include "plat_sensor_table.h"
+#include "plat_i2c.h"
+#include "sensor.h"
+
 LOG_MODULE_REGISTER(plat_hook);
+
+struct k_mutex i2c_1_PCA9546a_mutex;
+struct k_mutex i2c_2_PCA9546a_mutex;
+struct k_mutex i2c_6_pca9546a_mutex;
+struct k_mutex i2c_7_PCA9546a_mutex;
+struct k_mutex i2c_8_PCA9546a_mutex;
+struct k_mutex i2c_9_PCA9546a_mutex;
 
 /**************************************************************************************************
  * INIT ARGS
@@ -138,7 +147,36 @@ mux_config bus_9_PCA9546A_configs[] = {
 /**************************************************************************************************
  *  PRE-HOOK/POST-HOOK FUNC
  **************************************************************************************************/
+struct k_mutex *get_i2c_mux_mutex(uint8_t i2c_bus)
+{
+	struct k_mutex *mutex = NULL;
 
+	switch (i2c_bus) {
+	case I2C_BUS1:
+		mutex = &i2c_1_PCA9546a_mutex;
+		break;
+	case I2C_BUS2:
+		mutex = &i2c_2_PCA9546a_mutex;
+		break;
+	case I2C_BUS6:
+		mutex = &i2c_6_pca9546a_mutex;
+		break;
+	case I2C_BUS7:
+		mutex = &i2c_7_PCA9546a_mutex;
+		break;
+	case I2C_BUS8:
+		mutex = &i2c_8_PCA9546a_mutex;
+		break;
+	case I2C_BUS9:
+		mutex = &i2c_9_PCA9546a_mutex;
+		break;
+	default:
+		LOG_ERR("No support for i2c bus %d mutex", i2c_bus);
+		break;
+	}
+
+	return mutex;
+}
 bool pre_PCA9546A_read(sensor_cfg *cfg, void *args)
 {
 	CHECK_NULL_ARG_WITH_RETURN(cfg, false);
